Refactor C code generation
