+++ examples/Imperative.hs
@@ -6,6 +6,7 @@ import Control.Monad.Operational.Compositional
 import Language.C.Monad
 import Language.Embedded.Imperative
 import Language.Embedded.Expr
+import Language.Embedded.Backend.C
